Renaming reconnectionnotifier to resubscribe-notifier
