+++ test-unit/unit/utils/resubscribe-notifierSpec.js
